create idea str utility
