+++ src/a17_idea_logic/idea_config.py
@@ -16,62 +16,6 @@ def get_idea_config_dict() -> dict:
     return open_json(config_file_dir(), get_idea_config_filename())
 
 
-def idea_number_str() -> str:
-    return "idea_number"
-
-
-def idea_category_str() -> str:
-    return "idea_category"
-
-
-def get_idea_categorys() -> set[str]:
-    return {"bud", "fisc", "pidgin"}
-
-
-def allowed_crud_str() -> str:
-    return "allowed_crud"
-
-
-def dimens_str() -> str:
-    return "dimens"
-
-
-def otx_key_str() -> str:
-    return "otx_key"
-
-
-def attributes_str() -> str:
-    return "attributes"
-
-
-def insert_one_time_str() -> str:
-    return "INSERT_ONE_TIME"
-
-
-def insert_mulitple_str() -> str:
-    return "INSERT_MULITPLE"
-
-
-def delete_insert_update_str() -> str:
-    return "DELETE_INSERT_UPDATE"
-
-
-def insert_update_str() -> str:
-    return "INSERT_UPDATE"
-
-
-def delete_insert_str() -> str:
-    return "DELETE_INSERT"
-
-
-def delete_update_str() -> str:
-    return "DELETE_UPDATE"
-
-
-def build_order_str() -> str:
-    return "build_order"
-
-
 def get_allowed_curds() -> set[str]:
     return {
         "INSERT_ONE_TIME",
