+++ src/a17_idea_logic/test_idea/test_idea_format_builder.py
@@ -7,12 +7,8 @@ from src.a06_bud_logic._utils.str_a06 import (
     gogo_want_str,
 )
 from src.a08_bud_atom_logic.atom_config import get_atom_config_args
-from src.a17_idea_logic.idea_config import (
-    get_idea_formats_dir,
-    get_idea_config_dict,
-    dimens_str,
-    attributes_str,
-)
+from src.a17_idea_logic._utils.str_a17 import attributes_str, dimens_str
+from src.a17_idea_logic.idea_config import get_idea_formats_dir, get_idea_config_dict
 from src.a06_bud_logic._utils.str_a06 import bud_itemunit_str
